requirements.txt
streamlit>=1.28.0
google-api-python-client>=2.100.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1
notion-client>=2.0.0
markdown-it-py>=3.0.0
jinja2>=3.1.2
pyyaml>=6.0
python-dotenv>=1.0.0
humanize>=4.8.0
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.0.0
flake8>=6.1.0
